requirements.txt
torch
torchvision
matplotlib
tqdm
pillow
numpy